Fix the warnings test
