+++ tests/test_validation_interface/test_warnings_work.py
@@ -7,19 +7,18 @@ from fiasko_bro import validate
 
 
 @pytest.fixture(scope="session")
-def long_file_3_spaces_repo_path():
-    repo_path = 'test_fixtures{}long_file_3_spaces_repo'.format(os.path.sep)
+def file_3_spaces_repo_path():
+    repo_path = 'test_fixtures{}file_3_spaces_repo'.format(os.path.sep)
     remove_repo(repo_path)
     initialize_repo(repo_path)
     yield repo_path
     remove_repo(repo_path)
 
 
-def test_warnings_show_up_after_fail(long_file_3_spaces_repo_path):
+def test_warnings_show_up_after_fail(file_3_spaces_repo_path):
     expected_output = [
-        ('pep8', '240 PEP8 violations'),
-        ('file_too_long', 'long_file_3_spaces.py'),
-        ('indent_not_four_spaces', 'long_file_3_spaces.py:16')
+        ('pep8', '43 PEP8 violations'),
+        ('indent_not_four_spaces', 'file_3_spaces.py:16')
     ]
-    output = validate(long_file_3_spaces_repo_path)
+    output = validate(file_3_spaces_repo_path)
     assert output == expected_output
